+++ src/driver/compute.rs
@@ -11,6 +11,7 @@ use {
     derive_builder::Builder,
     log::{trace, warn},
     std::{
+        collections::HashSet,
         ffi::CString,
         fmt::{Debug, Formatter},
         hash::{Hash, Hasher},
@@ -77,13 +78,16 @@ impl ComputePipeline {
 
         // Use SPIR-V reflection to get the types and counts of all descriptors
         let mut descriptor_bindings = shader.descriptor_bindings();
-        for (descriptor_info, _) in descriptor_bindings.values_mut() {
+        let mut bindless_descriptors = HashSet::new();
+        for (descriptor, (descriptor_info, _)) in descriptor_bindings.iter_mut() {
             if descriptor_info.binding_count() == 0 {
+                bindless_descriptors.insert(*descriptor);
                 descriptor_info.set_binding_count(info.bindless_descriptor_count);
             }
         }
 
-        let descriptor_info = PipelineDescriptorInfo::create(device, &descriptor_bindings)?;
+        let descriptor_info =
+            PipelineDescriptorInfo::create(device, &descriptor_bindings, &bindless_descriptors)?;
         let descriptor_set_layouts = descriptor_info
             .layouts
             .values()